trino and flink all working as expected
