+++ README.md
@@ -17,7 +17,15 @@ We have collected here the code snippets found in the book (and a few extras too
 ## Chapter 2: Installing Delta Lake
 ## Chapter 3: Using Delta Lake
 ## Chapter 4: Delta Sharing and Unity Catalog
+
 ## Chapter 5: Diving into the Delta Lake Ecosystem
+[Chapter 5 : Source Code](./ch05/)
+
+### Apache Flink
+* [Environment Setup](./ch05/flink/README.md)
+* [Java Application](./ch05/flink/dldg-flink-delta-app/README.md)
+* **Docker Containers**: [amd64](./ch05/flink/docker-compose-flink.yaml) , [arm64](./ch05/flink/docker-compose-arm64-flink.yaml) 
+
 ## Chapter 6: Maintaining your Delta Lake
 ## Chapter 7: Under the Sediments (Delta Lake Internals)
 ## Chapter 8: tbd
@@ -28,8 +36,3 @@ We have collected here the code snippets found in the book (and a few extras too
 ## Chapter 13: Design Patterns
 ## Chapter 14: Security and Governance
 ## Appendix
-
-## Running Chapter Contents from Root
-~~~
-docker compose -f ch05/docker-compose-arm64-flink.yaml up --remove-orphans
-~~~
