--- ch05/.env
@@ -1,4 +1,0 @@
-FLINK_PROPERTIES="jobmanager.rpc.address: flink-job-manager
-taskmanager.numberOfTaskSlots: 2
-blob.server.port: 6124
-"
