--- ch05/applications/flink/dldg-flink-delta-app/.gitignore
@@ -1,8 +1,0 @@
-.idea/
-target/
-
-*.iml
-dependency-reduced-pom.xml
-
-.DS_Store
-
