--- ch05/applications/flink/dldg-flink-delta-app/src/main/resources/delta-source-app.properties
@@ -1,30 +1,0 @@
-delta.source.tablePath=/opt/data/delta/
-delta.source.tableName=ecomm_v1_clickstream
-
-# set isBounded to false to turn on Continuous execution mode
-delta.source.isBounded=true
-
-# control which columns are read (projected)
-delta.source.columnNames.enabled=false
-delta.source.columnNames=event_time,event_type,brand,price
-
-# control the starting position for the table using the (long) version id from the transaction log
-delta.source.versionAsOf.enabled=false
-delta.source.versionAsOf=0
-
-# control the starting position for the table read using the timestamp
-# string (Delta will find the closest match in the transaction log)
-delta.source.timestampAsOf.enabled=false
-delta.source.timestampAsOf=2023-08-31
-
-# control the Delta Lake table (sink) location
-delta.sink.tablePath=/opt/data/delta/
-delta.sink.tableName=ecomm_v1_clickstream_lite
-delta.sink.mergeSchema=false
-delta.sink.partitionCol=event_type
-
-
-flink.state.backend.enabled=true
-# options: file://, hdfs://, s3://
-flink.default.fs.prefix=file://
-flink.checkpoint.storage.dir=/opt/data/checkpoints/delta-source-app
