+++ ch05/flink/.env
@@ -1,0 +1,4 @@
+FLINK_PROPERTIES="jobmanager.rpc.address: flink-job-manager
+taskmanager.numberOfTaskSlots: 2
+blob.server.port: 6124
+"
