+++ ch05/flink/dldg-flink-delta-app/.gitignore
@@ -1,0 +1,8 @@
+.idea/
+target/
+
+*.iml
+dependency-reduced-pom.xml
+
+.DS_Store
+
