+++ ch05/flink/dldg-flink-delta-app/src/main/resources/app.properties
@@ -1,0 +1,15 @@
+delta.sink.tableName=ecomm_v1_clickstream
+delta.sink.mergeSchema=false
+delta.sink.partitionCols=event_type
+delta.sink.tablePath=/opt/data/delta/
+
+flink.state.backend.enabled=true
+# options: file://, hdfs://, s3://
+flink.default.fs.prefix=file://
+flink.checkpoint.storage.dir=/opt/data/checkpoints/
+
+kafka.brokers=PLAINTEXT://kafka-rp:29092
+kafka.topic=ecomm.v1.clickstream
+kafka.group.id=delta-dldg-1
+
+runtime.cleaner.delete.deltaTable=false
