+++ ch05/flink/dldg-flink-delta-app/src/test/resources/app.properties
@@ -1,0 +1,9 @@
+kafka.brokers=127.0.0.1:29092
+kafka.topic=ecomm.v1.clickstream
+kafka.group.id=delta-dldg-1
+
+delta.sink.tableName=ecomm_v1_clickstream
+delta.sink.mergeSchema=false
+delta.sink.tablePath=/opt/data/delta/
+
+runtime.cleaner.delete.deltaTable=false
