+++ ch05/trinodb/docker-compose-trino-arm64.yaml
@@ -7,80 +7,17 @@ services:
     hostname: trinodb
     container_name: trinodb
     volumes:
-      - $PWD/ch05/trinodb/etc/catalog:/etc/trino/catalog
+      - $PWD/etc/catalog/delta.properties:/etc/trino/catalog/delta.properties
+      - $PWD/conf:/etc/hadoop/conf/
     ports:
       - target: 8080
         published: 9090
         protocol: tcp
         mode: host
     environment:
-      - AWS_ACCESS_KEY_ID=minio
-      - AWS_SECRET_ACCESS_KEY=minio_admin
-      - AWS_DEFAULT_REGION=us-west-2
-    depends_on:
-      - minio
-      - metastore
-    networks:
-      - dldg
-  
-  metastore:
-    image: apache/hive:3.1.3
-    platform: linux/amd64
-    hostname: metastore
-    container_name: metastore
-    environment:
-      - SERVICE_NAME=metastore
-      - SERVICE_OPTS="-Dhive.metastore.uris=thrift://metastore:9083"
-      - IS_RESUME="true"
-    volumes:
-      - $PWD/data/hive/:/opt/hive/data/warehouse
-    expose:
-      - 9083
-      - 10000
-      - 10002
-    ports:
-      - target: 9083
-        published: 9083
-        protocol: tcp
-        mode: host
-      - target: 10000
-        published: 10000
-        protocol: tcp
-        mode: host
-      - target: 10002
-        published: 10002
-        protocol: tcp
-        mode: host
-    networks:
-      - dldg
-
-  minio:
-    image: minio/minio:RELEASE.2023-09-23T03-47-50Z
-    platform: linux/arm64
-    hostname: minio
-    container_name: minio
-    volumes:
-      - ${PWD}/data/minio/:/opt/data
-    environment:
-      MINIO_ROOT_USER: minio
-      MINIO_ROOT_PASSWORD: minio_admin
-    command: ["minio", "server", "--console-address", ":9001", "/opt/data"]
-    expose: 
-      - 9000
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:9000/minio/health/live"]
-      interval: 30s
-      timeout: 20s
-      retries: 3
-    ports:
-      - target: 9000
-        published: 9000
-        protocol: tcp
-        mode: host
-      - target: 9001
-        published: 9001
-        protocol: tcp
-        mode: host
+      - AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID:-minio}
+      - AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY:-minio_admin}
+      - AWS_DEFAULT_REGION=${AWS_DEFAULT_REGION:-us-west-1}
     networks:
       - dldg
