+++ ch05/trinodb/etc/catalog/delta.properties
@@ -1,0 +1,26 @@
+connector.name=delta_lake
+hive.metastore=thrift
+hive.metastore.uri=thrift://metastore:9083
+#hive.s3.aws-access-key=AKIATAXPTBZGM5HQGM7T
+#hive.s3.aws-secret-key=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
+#hive.s3.path-style-access=true
+#hive.s3.endpoint=http://minio:9000
+#hive.s3.aws-access-key=minio
+#hive.s3.aws-secret-key=minio_admin
+#hive.s3.ssl.enabled=false
+#hive.s3.aws-access-key=minio_trino
+#hive.s3.aws-secret-key=X6Kl9EgPvfuoQYyYyuyTiSx2nRPFc5Nhi5T7cZ9q
+#hive.s3.region=us-west-1
+delta.metadata.cache-ttl=5m
+delta.metadata.cache-size=10
+delta.compression-codec=SNAPPY
+delta.enable-non-concurrent-writes=true
+delta.default-checkpoint-writing-interval=10
+delta.hive-catalog-name=metastore
+delta.checkpoint-row-statistics-writing.enabled=true
+delta.table-statistics-enabled=true
+delta.extended-statistics.enabled=true
+delta.extended-statistics.collect-on-write=true
+delta.target-max-file-size=512MB
+delta.unique-table-location=true
+delta.vacuum.min-retention=7d
