--- ch05/trinodb/etc/catalog/deltalake.properties
@@ -1,21 +1,0 @@
-connector.name=delta_lake
-
-# https://trino.io/docs/current/connector/metastores.html#thrift-metastore-configuration-properties
-hive.metastore=thrift
-hive.metastore.uri=thrift://metastore:9083
-hive.metastore.authentication.type=NONE
-hive.metastore.thrift.use-spark-table-statistics-fallback=true
-
-delta.metadata.cache-ttl=5m
-delta.metadata.cache-size=10
-delta.compression-codec=SNAPPY
-delta.enable-non-concurrent-writes=true
-delta.default-checkpoint-writing-interval=10
-delta.hive-catalog-name=metastore_db
-delta.checkpoint-row-statistics-writing.enabled=true
-delta.table-statistics-enabled=true
-delta.extended-statistics.enabled=true
-delta.extended-statistics.collect-on-write=true
-delta.target-max-file-size=512MB
-delta.unique-table-location=true
-delta.vacuum.min-retention=7d
